move manager to its own package
