+++ plugin/manager/integration_test.go
@@ -1,4 +1,4 @@
-package plugin
+package manager
 
 import (
 	"io"
