+++ plugin/manager/testdata/main.go
